A couple of bugfixes.
ChefServerSlice::Application#load_all_files requires that cookbook file paths
look like cookbooks/name/segment/..., so my previous nested scheme was bogus.

And recipe templates are processed on the *client*, so that thing with
capturing the HTTP_HOST was also bogus. Now, a client has to be bootstrapped
with the proper server urls.